Backend logics flow

+++ ticketsystem/src/main/java/com/ticketsystem/ticketsystem/entity/Organization.java
@@ -45,7 +45,7 @@ public class Organization {
 
     @Column(name="org_phone")
     @NotBlank
-    private int orgPhone;
+    private Long orgPhone;
 
     @Column(name="industry_type")
     @NotBlank
